retrieve-image-tags/requirements.txt
PyGithub==1.58.0
requests==2.28.2
semver==2.13.0
